requirements.txt
numpy==2.4.3
scipy==1.17.1
scikit-learn==1.7.1
matplotlib==3.10.8
seaborn==0.13.2
Pillow==12.1.1
pandas==3.0.1
requests==2.32.3
tqdm==4.67.1
PyYAML==6.0.2
huggingface-hub==1.3.4
transformers==5.3.0
torch==2.5.1
mlx==0.31.1
mlx-lm==0.31.1
mlx-vlm==0.4.0
